Se sube que se genere la factura correspondiente al pago.

+++ Horizontes_Urbanos/src/main/java/CONTROLADOR/gestionar/GestionarVenta.java
@@ -39,7 +39,7 @@ public class GestionarVenta {
         persisVenta.editar(venta);
     }
     
-    public void borrar(int id){
+    public void borrar(Long id){
     
     }
